refactored the app tests to split them into controller tests
